REVERT THIS: always serve all-packages.json to 0.16 users
Since this is normally based on timestamp, it is likely that folks
upgrading to 0.16 will not get a new all-packages.json. If that happens
their constraints will not solve because they won’t have a viable
version of core.

This ensures that they are always getting new versions. It makes sense
to turn it off later though.