Match union tags when searching a package, fixing #112
Component.PackageSidebar.SearchDict's values are now pairs of strings.
The first is the name to display and the second is the name to link
to. In most cases these are the same, but union tags are displayed as
themselves but link to the name of their parent type. `toSearchDict`
learns how to find union tags from entries. Most of the other changes
in this module is propagating that logic down to where the links are
actually rendered.