cosmetic: clean up style
@mgold, please do not use multi-line anonymous functions. Also, please
do not use |> on a single line. It is always meant to be on a newline.
The big function in here was pretty unreadable with the parens in crazy
places.

Also, when submitting PRs to me, follow the style guide about adding
newlines. http://elm-lang.org/docs/style-guide#declarations This is
true in let expressions as well. Things come down a newline after a =
or a ->

It’s fine if you don’t follow these rules in your own code, but it does
not make sense for me to have to go and do this to every PR you do.

That said, nice work, this is a nice improvement!