Optimize the spacing of version dots on the timeline
The timeline is displayed at a fixed-width, but the variety of Elm
packages may have anywhere from one version to tens of versions.
This may lead to undesirable overlap of the dots that represent versions
in various common and not-so-common scenarios.

This change optimizes the spacing of dots on the timeline so that
the natural chronology is preserved as much as possible, while still
giving each version enough space. If there are simply too many dots
to be displayed in the allotted space, the algorithm will uniformly
squish everything to fit. This should only be necessary in pathological
cases (e.g. a package which has over 20 releases).

The key to the algorithm is representing the timeline as a list
of relative distances. For each version/dot, we store the distance
to the ADJACENT dot rather than the offset from the timeline origin.
This allows us to easily to make room for a dot by simply adjusting
one relative number without requiring a rippling change to each
downstream element.

The algorithm proceeds in multiple steps:
1) convert from the ProximityTree's representation of normalized
fractions measured from the origin to relative, adjacent distance
2) from left-to-right, adjust each element so that it has at least
its minimum space, stealing from downstream as necessary
3) after reaching the end, if the sum of the relative distances
is greater than 1, then we exceeded the fixed width and we need
to shrink
4) apply a "soft" shrink: proceeding from left-to-right COLLAPSE
as much empty space as you can without violating any minimum distances
until the overflow is consumed or we reach the end
5) if there is STILL overflow, apply a "hard" shrink: uniformly
scaling-down the entire timeline so that the last version fits
the fixed width
6) convert back to ProximityTree