Add libreirc/elm-crypto to native whitelist
elm-crypto wraps Web Cryptography API with Elm's Task interfaces.

LibreIRC is a working group who are trying to make an IRC client using
Elm. We found that it is necessary to wrap Web Cryptography API to
retrieve cryptographically secure pseudorandom number in Elm. So we
forked the blacktaxi/elm-random-secure and updated it to Elm 0.18.