only show package/version segments in the header
No more "help" or "ModuleName" in there.

The idea is that the header should focus on the author/project V.S.N
because the other information appears elsewhere a bunch of times. For
example, tabs like README, About, and module names will be bolded as
you navigate around. They also come with a really big header, so that
information already appears twice without the header.