update the /about page logic a bit
One major change is to store the whole (OneOrMore Release) in the model
and then derive the latest version and release dates from that. Since
this data is dependent on the same underlying date source, I think the
code is clearer when the model captures that and the view derives what
it needs.

Other than that, it is renaming Manifest to Outline and changing the
style a bit.