try to block crawler that seems to cause outages
I'm not certain yet, but it looks like both of the recent outages began with a request for `robots.txt` from a specific domain registered in Germany.

It looks like that company has many crawlers, and two specific crawlers can cause different failures:

- `ninja-crawler14.webmeup.com` can cause a short outage. Less than a second. There are a few examples of this in the last month.
- `ninja-crawler45.webmeup.com` caused the two long outages. It seems to cause some situation that is not recovered by `systemctl` restarts.

This is very weird to me, and perhaps I am misinterpreting things. I do not understand at all how a request for `robots.txt` could be a problem for the Haskell server. Maybe something weird in headers or something, but my logs are not detailed enough for me to know more about the past incidents.

They have some notes about how to have the `robots.txt` tell them to stop crawling at http://webmeup-crawler.com/ which is what I have changed here. I am not sure that'll be enough given that it seems that requesting `robots.txt` is the problem in the first place, but it seems like a good first step nonetheless.