attempt to skip Haskell when serving robots.txt
The current theory is that a specific crawler is requesting robots.txt in a way that causes issues for the Haskell binary. This change tries to serve robots.txt directly, skipping over the Haskell binary. Hopefully this will resolve the issue.

Looking through the logs, it does look like this crawler is relatively well behaved. It really does delay its requests to be once per three seconds. (Unlike some other IP addresses I saw in the logs!)