remove unneeded if logic from nginx.conf
The goal here is to handle legacy requests from Elm 0.18 and below.

Those binaries request /all-packages, but expect a different data
format than the current style. But looking back at the old code:

https://github.com/elm-lang/elm-package/blob/0.18.0/src/Catalog.hs#L70
https://github.com/elm-lang/elm-package/blob/0.18.0/src/Catalog.hs#L36
https://github.com/elm-lang/elm-package/blob/0.18.0/src/Manager.hs#L30

The hardcoded domain uses HTTP, so those requests would never go to
port 443 anyway. So it is not necessary to try to intercept the
elm-package-version query parameter here.